Renamed class to better indicate its purpose, added centralised close method and callback
